+++ sn_testnet/CHANGELOG.md
@@ -636,6 +636,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Other
 - update dependencies
 
+## [0.2.24](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.2.23...sn_testnet-v0.2.24) - 2023-07-19
+
+### Added
+- *(testnet)* enable the use of `CARGO_TARGET_DIR`
+
+## [0.2.25](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.2.24...sn_testnet-v0.2.25) - 2023-07-20
+
+### Other
+- update dependencies
+
+## [0.2.26](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.2.25...sn_testnet-v0.2.26) - 2023-07-20
+
+### Other
+- update dependencies
+
+## [0.2.27](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.2.26...sn_testnet-v0.2.27) - 2023-07-21
+
+### Other
+- update dependencies
+
 ## v0.1.0 (2023-03-16)
 
 <csr-id-4f04bd1a5d1c747bfc6b5d39824dd108f8546b7b/>